requirements.txt
amqp==5.3.1
audioread==3.0.1
av==14.1.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
ctranslate2==4.5.0
customtkinter==5.2.2
cv==1.0.0
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==25.1.24
fonttools==4.55.6
fsspec==2024.12.0
future==1.0.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiwer==3.1.0
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
onnxruntime==1.20.1
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-vlc==3.0.21203
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==3.4.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
